feat(dl): implement dynamic symbol loading module and Int64 handle unification

+++ native/src/sync_native.cpp
@@ -10,27 +10,27 @@
 
 extern "C" {
 
-double sync_sem_open(const char* name, int initial_value) {
+long long sync_sem_open(const char* name, int initial_value) {
     HANDLE h = CreateSemaphoreA(NULL, initial_value, 2147483647, name);
-    return (double)(size_t)h;
+    return (long long)(size_t)h;
 }
 
-int sync_sem_wait(double handle) {
+int sync_sem_wait(long long handle) {
     return (WaitForSingleObject((HANDLE)(size_t)handle, INFINITE) == WAIT_OBJECT_0) ? 0 : -1;
 }
 
-int sync_sem_post(double handle) {
+int sync_sem_post(long long handle) {
     return ReleaseSemaphore((HANDLE)(size_t)handle, 1, NULL) ? 0 : -1;
 }
 
-int sync_sem_trywait(double handle) {
+int sync_sem_trywait(long long handle) {
     DWORD res = WaitForSingleObject((HANDLE)(size_t)handle, 0);
     if (res == WAIT_OBJECT_0) return 0;
     if (res == WAIT_TIMEOUT) return -2;
     return -1;
 }
 
-void sync_sem_close(double handle) {
+void sync_sem_close(long long handle) {
     if (handle) CloseHandle((HANDLE)(size_t)handle);
 }
 
@@ -50,29 +50,29 @@ void sync_sem_unlink(const char* name) {
 
 extern "C" {
 
-double sync_sem_open(const char* name, int initial_value) {
+long long sync_sem_open(const char* name, int initial_value) {
     std::string posix_name = name;
     if (posix_name[0] != '/') posix_name = "/" + posix_name;
     
     sem_t* sem = sem_open(posix_name.c_str(), O_CREAT, 0666, initial_value);
-    return (sem == SEM_FAILED) ? 0 : (double)(size_t)sem;
+    return (sem == SEM_FAILED) ? 0 : (long long)(size_t)sem;
 }
 
-int sync_sem_wait(double handle) {
+int sync_sem_wait(long long handle) {
     return sem_wait((sem_t*)(size_t)handle);
 }
 
-int sync_sem_post(double handle) {
+int sync_sem_post(long long handle) {
     return sem_post((sem_t*)(size_t)handle);
 }
 
-int sync_sem_trywait(double handle) {
+int sync_sem_trywait(long long handle) {
     if (sem_trywait((sem_t*)(size_t)handle) == 0) return 0;
     if (errno == EAGAIN) return -2;
     return -1;
 }
 
-void sync_sem_close(double handle) {
+void sync_sem_close(long long handle) {
     if (handle) sem_close((sem_t*)(size_t)handle);
 }
